requirements.txt
streamlit
pinecone-client
sentence-transformers
openai>=1.54.4
python-dotenv
pandas
boto3
